--- src/middlewares/ErrorHandlerMiddleware.js
@@ -1,30 +1,0 @@
-import {AppError, BadRequestError, NotFoundError, ValidationError} from "../utils/AppError.ts";
-import {ResponseHandler} from "../utils/ResponseHandler.js";
-
-export const errorHandlerMiddleware = async (req, res, next) => {
-    try {
-        await next();
-    } catch (error) {
-        console.error({
-            message: error.message, stack: error.stack,
-        })
-
-        if (error instanceof ValidationError) {
-            return ResponseHandler.error(res, error.message, {statusCode: 400});
-        }
-
-        if (error instanceof NotFoundError) {
-            return ResponseHandler.error(res, error.message, {statusCode: 404});
-        }
-
-        if (error instanceof BadRequestError) {
-            return ResponseHandler.error(res, error.message, {statusCode: 400});
-        }
-
-        if (error instanceof AppError) {
-            return ResponseHandler.error(res, error.message, {statusCode: 400});
-        }
-
-        return ResponseHandler.error(res, '예상치 못한 오류 입니다.', {statusCode: 500});
-    }
-};
